+++ CHANGELOG.md
@@ -37,6 +37,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+### Improvements
+
+* [#43](https://github.com/babylonlabs-io/covenant-emulator/pull/43) Test importing
+private key to the cosmos keyring
+
 ## v0.9.0
 
 * [#33](https://github.com/babylonlabs-io/covenant-emulator/pull/33) Add remote